backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.1
earthengine-api==0.1.418
python-multipart==0.0.9
torch>=2.9.0
torchvision>=0.20.0
Pillow>=11.0.0
python-dotenv==1.0.1
google-genai>=1.0.0
